+++ .env.local
@@ -1,3 +1,3 @@
 #NOTE: only variables starting with "REACT_APP_" will be read
-#REACT_APP_DOMAIN = http://localhost:8000
-REACT_APP_DOMAIN = http://api.omoulearning.net
+REACT_APP_DOMAIN = http://localhost:8000
+#REACT_APP_DOMAIN = http://api.omoulearning.net
