+++ src/components/FeatureViews/Accounts/ProfileCard.js
@@ -1,14 +1,27 @@
 import {connect} from "react-redux";
 import React, {Component} from "react";
+import BackButton from "../../BackButton";
 import Grid from "@material-ui/core/Grid";
-import {Card, Typography} from "@material-ui/core";
+import {Card, Paper, Typography} from "@material-ui/core";
+import Tabs from "@material-ui/core/Tabs";
+import Tab from "@material-ui/core/Tab";
+import ListView from "@material-ui/icons/ViewList";
+import CardView from "@material-ui/icons/ViewModule";
+import TableHead from "@material-ui/core/TableHead";
+import TableRow from "@material-ui/core/TableRow";
+import TableCell from "@material-ui/core/TableCell";
+import Table from "@material-ui/core/Table";
+import TableBody from "@material-ui/core/TableBody";
 import CardContent from "@material-ui/core/CardContent";
+import Button from "@material-ui/core/Button";
 import Chip from "@material-ui/core/Chip";
+import CardActions from "@material-ui/core/CardActions";
 import {withRouter} from "react-router-dom";
 import EmailIcon from "@material-ui/icons/EmailOutlined";
 import PhoneIcon from "@material-ui/icons/PhoneOutlined";
 import Hidden from "@material-ui/core/es/Hidden/Hidden";
-import {addDashes, stringToColor} from "./accountUtils";
+import { stringToColor } from "./accountUtils";
+import { addDashes } from "./accountUtils";
 
 import {ReactComponent as IDIcon} from "../../identifier.svg";
 import './Accounts.scss';
