+++ src/components/FeatureViews/Notes/Notes.js
@@ -1,6 +1,5 @@
 import React, {useCallback, useEffect, useMemo, useState} from "react";
 import * as userActions from "../../../actions/userActions";
-import * as hooks from "actions/hooks";
 import {useDispatch, useSelector} from "react-redux";
 import {bindActionCreators} from "redux";
 import PropTypes from "prop-types";
@@ -11,7 +10,6 @@ import DoneIcon from "@material-ui/icons/CheckCircleOutlined";
 import EditIcon from "@material-ui/icons/EditOutlined";
 import Grid from "@material-ui/core/Grid";
 import InputBase from "@material-ui/core/InputBase";
-import Loading from "components/Loading";
 import Modal from "@material-ui/core/Modal";
 import NotificationIcon from "@material-ui/icons/NotificationImportant";
 import Paper from "@material-ui/core/Paper";
@@ -22,68 +20,48 @@ import "../Accounts/TabComponents/TabComponents.scss";
 import {GET, PATCH, POST} from "../../../actions/actionTypes";
 import {REQUEST_STARTED} from "../../../actions/apiActions";
 
-const numericDateString = (date) =>
-    (new Date(date)).toLocaleTimeString("en-US", {
+const numericDateString = (date) => {
+    const DateObject = new Date(date);
+    return DateObject.toLocaleTimeString("en-US", {
         "day": "numeric",
         "hour": "2-digit",
         "minute": "2-digit",
         "month": "numeric",
     });
+};
 
-const Notes = ({ownerType, ownerID}) => {
+const Notes = ({userRole, userID}) => {
     const dispatch = useDispatch();
     const api = useMemo(() => bindActionCreators(userActions, dispatch), [dispatch]);
-    const notes = useSelector(({Users, Course, Enrollments}) => {
-        switch (ownerType) {
+    const user = useSelector(({Users, Course}) => {
+        switch (userRole) {
             case "student":
-                return Users.StudentList[ownerID].notes;
+                return Users.StudentList[userID];
             case "parent":
-                return Users.ParentList[ownerID].notes;
+                return Users.ParentList[userID];
             case "instructor":
-                return Users.InstructorList[ownerID].notes;
+                return Users.InstructorList[userID];
             case "receptionist":
-                return Users.ReceptionistList[ownerID].notes;
+                return Users.ReceptionistList[userID];
             case "course":
-                return Course.NewCourseList[ownerID].notes;
-            case "enrollment":
-                return Enrollments[ownerID.studentID][ownerID.courseID].notes;
+                return Course.NewCourseList[userID];
             default:
-                return null;
+                return -1;
         }
     });
 
-    const getRequestStatus = useSelector(({RequestStatus}) => {
-        switch (ownerType) {
-            case "course":
-                return RequestStatus.courseNote[GET][ownerID];
-            case "enrollment":
-                return RequestStatus.enrollmentNote[GET][ownerID.enrollmentID];
-            default:
-                return RequestStatus.accountNote[GET][ownerID];
-        }
-    });
-
-    const postRequestStatus = useSelector(({RequestStatus}) => {
-        switch (ownerType) {
-            case "course":
-                return RequestStatus.courseNote[POST];
-            case "enrollment":
-                return RequestStatus.enrollmentNote[POST];
-            default:
-                return RequestStatus.accountNote[POST];
-        }
-    });
-
-    const patchRequestStatus = useSelector(({RequestStatus}) => {
-        switch (ownerType) {
-            case "course":
-                return RequestStatus.courseNote[PATCH][ownerID];
-            case "enrollment":
-                return RequestStatus.enrollmentNote[PATCH][ownerID.enrollmentID];
-            default:
-                return RequestStatus.accountNote[PATCH][ownerID];
-        }
-    });
+    const getRequestStatus = useSelector(({RequestStatus}) =>
+        userRole === "course"
+            ? RequestStatus.courseNote[GET][userID]
+            : RequestStatus.note[GET][userID]);
+    const postRequestStatus = useSelector(({RequestStatus}) =>
+        userRole === "course"
+            ? RequestStatus.courseNote[POST]
+            : RequestStatus.note[POST]);
+    const patchRequestStatus = useSelector(({RequestStatus}) =>
+        userRole === "course"
+            ? RequestStatus.courseNote[PATCH][userID]
+            : RequestStatus.note[PATCH][userID]);
 
     const [alert, setAlert] = useState(false);
     const [noteBody, setNoteBody] = useState("");
@@ -93,16 +71,6 @@ const Notes = ({ownerType, ownerID}) => {
     const [submitting, setSubmitting] = useState(false);
     const [isPost, setIsPost] = useState(false);
     const [error, setError] = useState(false);
-
-    useEffect(() => {
-        if (ownerType === "course") {
-            api.fetchCourseNotes(ownerID);
-        } else if (ownerType === "enrollment") {
-            api.fetchEnrollmentNotes(ownerID.enrollmentID, ownerID.studentID, ownerID.courseID);
-        } else {
-            api.fetchAccountNotes(ownerID, ownerType);
-        }
-    }, [api, ownerID, ownerType]);
 
     const openNewNote = useCallback(() => {
         setAlert(true);
@@ -142,68 +110,46 @@ const Notes = ({ownerType, ownerID}) => {
     };
 
     const saveNote = useCallback(() => {
-        switch (ownerType) {
-            case "enrollment": {
-                const note = {
-                    "body": noteBody,
-                    "complete": false,
-                    "important": notification,
-                    "title": noteTitle,
-                    "enrollment": ownerID.enrollmentID,
-                };
-                if (editID) {
-                    api.patchEnrollmentNote(editID, note, ownerID.enrollmentID, ownerID.studentID, ownerID.courseID);
-                    setIsPost(false);
-                } else {
-                    api.postEnrollmentNote(note, ownerID.enrollmentID, ownerID.studentID, ownerID.courseID);
-                    setIsPost(true);
-                }
-                setSubmitting(true);
-                break;
+        if (userRole === "course") {
+            const note = {
+                "body": noteBody,
+                "complete": false,
+                "important": notification,
+                "title": noteTitle,
+                "course": userID,
+            };
+            if (editID) {
+                api.patchCourseNote(editID, note, userRole);
+                setIsPost(false);
+            } else {
+                api.postCourseNote(note, userRole);
+                setIsPost(true);
             }
-            case "course": {
-                const note = {
-                    "body": noteBody,
-                    "complete": false,
-                    "important": notification,
-                    "title": noteTitle,
-                    "course": ownerID,
-                };
-                if (editID) {
-                    api.patchCourseNote(editID, note, ownerType, ownerID);
-                    setIsPost(false);
-                } else {
-                    api.postCourseNote(note, ownerType);
-                    setIsPost(true);
-                }
-                setSubmitting(true);
-                break;
+            setSubmitting(true);
+        } else {
+            const note = {
+                "body": noteBody,
+                "complete": false,
+                "important": notification,
+                "title": noteTitle,
+                "user": userID,
+            };
+            if (editID) {
+                api.patchNote(editID, note, userRole);
+                setIsPost(false);
+            } else {
+                api.postNote(note, userRole);
+                setIsPost(true);
             }
-            default: {
-                const note = {
-                    "body": noteBody,
-                    "complete": false,
-                    "important": notification,
-                    "title": noteTitle,
-                    "user": ownerID,
-                };
-                if (editID) {
-                    api.patchAccountNote(editID, note, ownerType, ownerID);
-                    setIsPost(false);
-                } else {
-                    api.postAccountNote(note, ownerType);
-                    setIsPost(true);
-                }
-                setSubmitting(true);
-            }
+            setSubmitting(true);
         }
-    }, [api, noteBody, notification, noteTitle, ownerID, ownerType, editID]);
+    }, [api, noteBody, notification, noteTitle, userID, userRole, editID]);
 
-    if (hooks.isLoading(getRequestStatus) && (!notes || Object.entries(notes).length === 0)) {
-        return <Loading />;
+    if (!getRequestStatus || getRequestStatus === REQUEST_STARTED) {
+        return "Loading notes...";
     }
 
-    if (hooks.isFail(getRequestStatus) && (!notes || Object.entries(notes).length === 0)) {
+    if (getRequestStatus < 200 || getRequestStatus >= 300) {
         return "Error loading notes!";
     }
 
@@ -243,6 +189,7 @@ const Notes = ({ownerType, ownerID}) => {
                             className="textfield"
                             id="standard-name"
                             label="Subject"
+                            margin="normal"
                             onChange={handleTitleUpdate}
                             value={noteTitle} />
                         <NotificationIcon
@@ -299,7 +246,7 @@ const Notes = ({ownerType, ownerID}) => {
                     </Typography>
                 </div>
             </Grid>
-            {notes && Object.values(notes).map((note) => (
+            {user.notes && Object.values(user.notes).map((note) => (
                 <Grid
                     item
                     key={note.id || note.body}
@@ -337,32 +284,11 @@ const Notes = ({ownerType, ownerID}) => {
 };
 
 Notes.propTypes = {
-    "ownerID": PropTypes.oneOfType([
+    "userID": PropTypes.oneOfType([
         PropTypes.string,
         PropTypes.number,
-        PropTypes.shape({
-            "courseID": PropTypes.oneOfType([
-                PropTypes.string,
-                PropTypes.number,
-            ]),
-            "enrollmentID": PropTypes.oneOfType([
-                PropTypes.string,
-                PropTypes.number,
-            ]),
-            "studentID": PropTypes.oneOfType([
-                PropTypes.string,
-                PropTypes.number,
-            ]),
-        }),
     ]).isRequired,
-    "ownerType": PropTypes.oneOf([
-        "course",
-        "enrollment",
-        "instructor",
-        "parent",
-        "receptionist",
-        "student",
-    ]).isRequired,
+    "userRole": PropTypes.string.isRequired,
 };
 
 export default Notes;
