+++ src/components/FeatureViews/Registration/RegistrationActions.js
@@ -1,13 +1,19 @@
 import PropTypes from "prop-types";
-import React, {useState} from "react";
+import React, { useState } from "react";
+
 // Material UI Imports
 import Grid from "@material-ui/core/Grid";
 import Button from "@material-ui/core/Button";
 import NewUser from "@material-ui/icons/PersonAdd";
-import {NavLink} from "react-router-dom";
+import NewTutor from "@material-ui/icons/Group";
+import NewCourse from "@material-ui/icons/School";
+import { NavLink } from "react-router-dom";
 import Menu from "@material-ui/core/Menu";
-import {withStyles} from "@material-ui/core/styles";
+import MenuItem from "@material-ui/core/MenuItem";
+import { withStyles } from "@material-ui/core/styles";
 import "./registration.scss";
+import ListItemText from "@material-ui/core/ListItemText";
+import AssignmentIcon from "@material-ui/icons/Assignment";
 import SetParent from "./SetParentRegistration";
 import {bindActionCreators} from "redux";
 import * as registrationActions from "../../../actions/registrationActions";
@@ -62,9 +68,7 @@ function RegistrationActions(props) {
             </Grid>
             <Grid item md={8}>
                 {
-                    (props.registration.CurrentParent &&
-                    props.registration.CurrentParent !== "none") &&
-                    <ParentRegistrationActions/>
+                    props.registration.CurrentParent && <ParentRegistrationActions/>
                 }
             </Grid>
             <SetParent/>
